+++ library/src/main/kotlin/org/jraf/android/kprefs/PreferenceFlowProperty.kt
@@ -29,58 +29,128 @@ package org.jraf.android.kprefs
 import android.content.SharedPreferences
 import android.content.SharedPreferences.OnSharedPreferenceChangeListener
 import kotlinx.coroutines.ExperimentalCoroutinesApi
-import kotlinx.coroutines.channels.awaitClose
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.callbackFlow
+import kotlinx.coroutines.InternalCoroutinesApi
+import kotlinx.coroutines.flow.FlowCollector
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
 import org.jraf.android.kprefs.Prefs.Companion.getKey
 import kotlin.properties.ReadOnlyProperty
 import kotlin.reflect.KProperty
 
-private fun <T> nonNullPreferenceFlow(
-    sharedPreferences: SharedPreferences,
-    key: String,
-    default: T,
-    getter: SharedPreferences.(String, T) -> T?
-): Flow<T> = callbackFlow {
-    fun offer() = offer(if (!sharedPreferences.contains(key)) default else sharedPreferences.getter(key, default)!!)
-    offer()
+private abstract class WrappedMutableStateFlow<T> : MutableStateFlow<T> {
+    private val wrapped by lazy { MutableStateFlow(getInitialValue()) }
 
-    val listener = OnSharedPreferenceChangeListener { _, _ ->
-        offer()
-    }
-    sharedPreferences.registerOnSharedPreferenceChangeListener(listener)
-    awaitClose {
-        sharedPreferences.unregisterOnSharedPreferenceChangeListener(listener)
-    }
+    override val replayCache: List<T> get() = wrapped.replayCache
+
+    override val subscriptionCount: StateFlow<Int> get() = wrapped.subscriptionCount
+
+    override var value: T
+        get() = wrapped.value
+        set(value) {
+            wrapped.value = value
+        }
+
+    @InternalCoroutinesApi
+    override suspend fun collect(collector: FlowCollector<T>) = wrapped.collect(collector)
+
+    override fun compareAndSet(expect: T, update: T): Boolean = wrapped.compareAndSet(expect, update)
+
+    override suspend fun emit(value: T) = wrapped.emit(value)
+
+    @ExperimentalCoroutinesApi
+    override fun resetReplayCache() = wrapped.resetReplayCache()
+
+    override fun tryEmit(value: T): Boolean = wrapped.tryEmit(value)
+
+    abstract fun getInitialValue(): T
 }
 
-private fun <T> nullablePreferenceFlow(
-    sharedPreferences: SharedPreferences,
-    key: String,
-    default: T,
-    getter: SharedPreferences.(String, T) -> T?
-): Flow<T?> = callbackFlow {
-    fun offer() = offer(if (!sharedPreferences.contains(key)) null else sharedPreferences.getter(key, default))
-    offer()
+private class NonNullPreferenceFlow<T : Any>(
+    private val sharedPreferences: SharedPreferences,
+    private val key: String,
+    private val default: T,
+    private val getter: SharedPreferences.(String, T) -> T?,
+    private val setter: SharedPreferences.Editor.(String, T) -> SharedPreferences.Editor,
+) : WrappedMutableStateFlow<T>() {
 
-    val listener = OnSharedPreferenceChangeListener { _, _ ->
-        offer()
+    private val listener = OnSharedPreferenceChangeListener { _, key ->
+        if (this.key == key) super.value = getPreferenceValue()
     }
-    sharedPreferences.registerOnSharedPreferenceChangeListener(listener)
-    awaitClose {
-        sharedPreferences.unregisterOnSharedPreferenceChangeListener(listener)
+
+    init {
+        sharedPreferences.registerOnSharedPreferenceChangeListener(listener)
     }
+
+    private fun getPreferenceValue(): T {
+        return if (!sharedPreferences.contains(key)) default else sharedPreferences.getter(this.key, default)!!
+    }
+
+    private fun updateSharedPreference(value: T) {
+        sharedPreferences.edit().apply {
+            setter(key, value)
+            apply()
+        }
+    }
+
+    override var value: T
+        get() = super.value
+        set(value) {
+            updateSharedPreference(value)
+        }
+
+    override fun getInitialValue() = getPreferenceValue()
+}
+
+private class NullablePreferenceFlow<T>(
+    private val sharedPreferences: SharedPreferences,
+    private val key: String,
+    private val default: T,
+    private val getter: SharedPreferences.(String, T) -> T?,
+    private val setter: SharedPreferences.Editor.(String, T) -> SharedPreferences.Editor
+) : WrappedMutableStateFlow<T?>() {
+
+    private val listener = OnSharedPreferenceChangeListener { _, key ->
+        if (this.key == key) super.value = getPreferenceValue()
+    }
+
+    init {
+        sharedPreferences.registerOnSharedPreferenceChangeListener(listener)
+    }
+
+    private fun getPreferenceValue(): T? {
+        return if (!sharedPreferences.contains(key)) null else sharedPreferences.getter(this.key, default)
+    }
+
+    private fun updateSharedPreference(value: T?) {
+        sharedPreferences.edit().apply {
+            if (value == null) {
+                remove(key)
+            } else {
+                setter(key, value)
+            }
+            apply()
+        }
+    }
+
+    override var value: T?
+        get() = super.value
+        set(value) {
+            updateSharedPreference(value)
+        }
+
+    override fun getInitialValue() = getPreferenceValue()
 }
 
 
-internal class NonNullPreferenceFlowProperty<T>(
+internal class NonNullPreferenceFlowProperty<T : Any>(
     private val sharedPreferences: SharedPreferences,
     private val key: String?,
     private val default: T,
-    private val getter: SharedPreferences.(String, T) -> T?
-) : ReadOnlyProperty<Any, Flow<T>> {
-    override fun getValue(thisRef: Any, property: KProperty<*>): Flow<T> {
-        return nonNullPreferenceFlow(sharedPreferences, getKey(property, key), default, getter)
+    private val getter: SharedPreferences.(String, T) -> T?,
+    private val setter: SharedPreferences.Editor.(String, T) -> SharedPreferences.Editor
+) : ReadOnlyProperty<Any, MutableStateFlow<T>> {
+    override fun getValue(thisRef: Any, property: KProperty<*>): MutableStateFlow<T> {
+        return NonNullPreferenceFlow(sharedPreferences, getKey(property, key), default, getter, setter)
     }
 }
 
@@ -88,9 +158,10 @@ internal class NullablePreferenceFlowProperty<T>(
     private val sharedPreferences: SharedPreferences,
     private val key: String?,
     private val default: T,
-    private val getter: SharedPreferences.(String, T) -> T?
-) : ReadOnlyProperty<Any, Flow<T?>> {
-    override fun getValue(thisRef: Any, property: KProperty<*>): Flow<T?> {
-        return nullablePreferenceFlow(sharedPreferences, getKey(property, key), default, getter)
+    private val getter: SharedPreferences.(String, T) -> T?,
+    private val setter: SharedPreferences.Editor.(String, T) -> SharedPreferences.Editor
+) : ReadOnlyProperty<Any, MutableStateFlow<T?>> {
+    override fun getValue(thisRef: Any, property: KProperty<*>): MutableStateFlow<T?> {
+        return NullablePreferenceFlow(sharedPreferences, getKey(property, key), default, getter, setter)
     }
 }