+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v1.5.0 (2020-10-31)
 - Update dependencies to latest versions
+- [Breaking change] Use of `Key` for all types instead of only for String (for consistency)
 
 ## v1.4.0 (2020-10-04)
 LiveData wrappers are now MutableLiveData.