+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## v1.6.0 (2020-10-31)
+Flow wrappers are now MutableStateFlow.
+
 ## v1.5.0 (2020-10-31)
 - Update dependencies to latest versions
 - [Breaking change] Use of `Key` for all types instead of only for String (for consistency)